+++ services/api/src/routes/__tests__/auth.js
@@ -1,6 +1,5 @@
 const jwt = require('jsonwebtoken');
 const { createTemporaryToken, generateTokenId } = require('../../utils/tokens');
-const { generateSecret, generateToken } = require('../../utils/mfa');
 const { setupDb, teardownDb, request, createUser } = require('../../utils/testing');
 const { mockTime, unmockTime, advanceTime } = require('../../utils/testing/time');
 const { User } = require('../../models');
@@ -59,7 +58,6 @@ describe('/1/auth', () => {
       expect(response.status).toBe(401);
 
       advanceTime(60 * 1000);
-
       response = await request('POST', '/1/auth/login', { email: user.email, password });
       expect(response.status).toBe(200);
 
@@ -149,148 +147,6 @@ describe('/1/auth', () => {
 
       response = await request('POST', '/1/auth/register', { firstName, lastName, email, password });
       expect(response.status).toBe(400);
-    });
-  });
-
-  describe('POST /mfa/verify', () => {
-    it('should verify mfa token', async () => {
-      const user = await createUser({
-        mfaMethod: 'otp',
-        tempTokenId: generateTokenId(),
-      });
-      const secret = generateSecret({
-        name: 'APP_TEST',
-        account: user,
-      });
-      user.mfaSecret = secret.secret;
-      await user.save();
-      const code = generateToken(secret.secret);
-
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      const response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      const { payload } = jwt.decode(response.body.data.token, { complete: true });
-      expect(payload).toHaveProperty('kid', 'user');
-      expect(payload).toHaveProperty('type', 'user');
-    });
-
-    it('should verify backup code', async () => {
-      const backupCode = '12345-16123';
-      const user = await createUser({
-        backupCodes: [backupCode],
-      });
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      const response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code: backupCode },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      const { payload } = jwt.decode(response.body.data.token, { complete: true });
-      expect(payload).toHaveProperty('kid', 'user');
-      expect(payload).toHaveProperty('type', 'user');
-    });
-
-    it('should failed with bad code', async () => {
-      const user = await createUser({
-        mfaMethod: 'otp',
-        tempTokenId: generateTokenId(),
-      });
-      const secret = generateSecret({
-        name: 'APP_TEST',
-        account: user,
-      });
-      user.mfaSecret = secret.secret;
-      await user.save();
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      const response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code: 'bad code' },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      expect(response.status).toBe(400);
-      expect(response.body.error.message).toBe('Not a valid code');
-    });
-
-    it('should failed with bad backup code', async () => {
-      const badBackupCode = '12345-16123';
-      const user = await createUser({});
-      await user.save();
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      const response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code: badBackupCode },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      expect(response.status).toBe(400);
-      expect(response.body.error.message).toBe('Not a valid code');
-    });
-
-    it('should increase login attempts', async () => {
-      const goodBackupCode = '92345-14812';
-      const user = await createUser({
-        backupCodes: [goodBackupCode],
-        loginAttempts: 1,
-      });
-      await user.save();
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      let response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code: '123123' },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      const dbUser = await User.findById(user.id);
-      expect(dbUser.loginAttempts).toBe(2);
-      expect(response.status).toBe(400);
-    });
-
-    it('should block user if limit is reached', async () => {
-      const goodBackupCode = '92345-14812';
-      const user = await createUser({
-        backupCodes: [goodBackupCode],
-        loginAttempts: 10,
-        lastLoginAttemptAt: new Date(),
-      });
-      await user.save();
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      let response = await request(
-        'POST',
-        '/1/auth/mfa/verify',
-        { code: '123123' },
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      expect(response.status).toBe(401);
-      expect(response.body.error.message).toBe('Too many attempts');
-    });
-  });
-
-  describe('POST /mfa/send-token', () => {
-    it('should trigger a token being sent', async () => {
-      const tokenId = generateTokenId();
-      const user = await createUser({
-        mfaMethod: 'sms',
-        mfaSecret: generateSecret({
-          name: 'APP_TEST',
-          account: '',
-        }).secret,
-        mfaPhoneNumber: '123123123',
-        tempTokenId: tokenId,
-      });
-      const token = createTemporaryToken({ type: 'mfa', sub: user.id, jti: user.tempTokenId });
-      let response = await request(
-        'POST',
-        '/1/auth/mfa/send-token',
-        {},
-        { headers: { Authorization: `Bearer ${token}` } }
-      );
-      expect(response.status).toBe(204);
     });
   });
