+++ services/api/src/routes/__tests__/users.js
@@ -260,191 +260,4 @@ describe('/1/users', () => {
       expect(response.status).toBe(401);
     });
   });
-
-  describe('POST /me/mfa/config', () => {
-    it('should allow mfa config (otp)', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(),
-      });
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/config`,
-        {
-          method: 'otp',
-        },
-        { user }
-      );
-      expect(response.status).toBe(200);
-      expect(response.body.data.secret).toBeDefined();
-    });
-
-    it('should allow mfa config (sms)', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(),
-      });
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/config`,
-        {
-          method: 'sms',
-          phoneNumber: '1231231',
-        },
-        { user }
-      );
-      expect(response.status).toBe(200);
-      expect(response.body.data.secret).toBeDefined();
-    });
-
-    it('should fail if you access has not be validated', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(0),
-      });
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/config`,
-        {
-          method: 'sms',
-          phoneNumber: '1231231',
-        },
-        { user }
-      );
-      expect(response.status).toBe(403);
-    });
-  });
-
-  describe('POST /me/mfa/enable', () => {
-    it('should allow mfa to be enabled (otp)', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(),
-      });
-      const { secret } = generateSecret({
-        name: 'APP_TEST',
-        account: user,
-      });
-      const code = generateToken(secret);
-
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/enable`,
-        {
-          code,
-          secret,
-          method: 'otp',
-          backupCodes: ['burger'],
-        },
-        { user }
-      );
-      expect(response.status).toBe(204);
-      const dbUser = await User.findById(user.id);
-      expect(dbUser.mfaSecret).toBe(secret);
-      expect(dbUser.mfaMethod).toBe('otp');
-      expect(dbUser.backupCodes[0]).toBe('burger');
-    });
-
-    it('should allow mfa to be enabled (sms)', async () => {
-      const phoneNumber = '123123123';
-      const user = await createUser({
-        accessConfirmedAt: new Date(),
-      });
-      const { secret } = generateSecret({
-        name: 'APP_TEST',
-        account: user,
-      });
-      const code = generateToken(secret);
-
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/enable`,
-        {
-          phoneNumber,
-          code,
-          secret,
-          method: 'sms',
-          backupCodes: ['burger'],
-        },
-        { user }
-      );
-      expect(response.status).toBe(204);
-      const dbUser = await User.findById(user.id);
-      expect(dbUser.mfaSecret).toBe(secret);
-      expect(dbUser.mfaMethod).toBe('sms');
-      expect(dbUser.backupCodes[0]).toBe('burger');
-      expect(dbUser.mfaPhoneNumber).toBe(phoneNumber);
-    });
-
-    it('should fail with a bad code', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(),
-      });
-      const { secret } = generateSecret({
-        name: 'APP_TEST',
-        account: user,
-      });
-
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/enable`,
-        {
-          code: 'bad',
-          secret,
-          method: 'otp',
-          backupCodes: ['burger'],
-        },
-        { user }
-      );
-
-      expect(response.status).toBe(400);
-      expect(response.body.error.message).toBe('Not a valid code');
-    });
-
-    it('should fail if you access has not be validated', async () => {
-      const user = await createUser({
-        accessConfirmedAt: new Date(0),
-      });
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/enable`,
-        {
-          code: 'bad',
-          secret: '1213123',
-          method: 'otp',
-          backupCodes: ['burger'],
-        },
-        { user }
-      );
-      expect(response.status).toBe(403);
-    });
-  });
-
-  describe('POST /me/mfa/generate-codes', () => {
-    it('should generate new codes', async () => {
-      const user = await createUser({});
-      const response = await request('POST', `/1/users/me/mfa/generate-codes`, {}, { user });
-      expect(response.status).toBe(200);
-      expect(response.body.data[0]).toBeDefined();
-    });
-  });
-
-  describe('POST /me/mfa/verify', () => {
-    it('should verify a code', async () => {
-      const user = await createUser({});
-      const { secret } = generateSecret({
-        name: 'APP_TEST',
-        account: '12213',
-      });
-      const code = generateToken(secret);
-
-      const response = await request(
-        'POST',
-        `/1/users/me/mfa/verify`,
-        {
-          code,
-          secret: secret,
-          method: 'sms',
-        },
-        { user }
-      );
-      expect(response.status).toBe(204);
-    });
-  });
 });
